Update documentation to match reality
Currently PySPLIT looks at the last character of a filename (including extension, if any) to find the file number.  Weekly files end with the number of the week in that month (1 through 4 or 5); monthly files should end with 1 and meteo_bookends should be changed to `([1], [1])`.  The documentation previously made it seem that files with the appropriate number but other endings would be found successfully and this is not true at this time.